Updates for NT4 (MSVC5++). Removed external include/library paths from projects.  External paths should be set via Tools | Options | Directories.  This allows each developer the freedom to install external libraries where they desire. Used libdb.lib instead of libdbs.lib to avoid thread conflicts. Added hs_regex.lib to library input.  We require some form of regex, this library works (and is relatively easy for the user to install). Removed a little lint which MCVC5 detected. Need to sort out single-threaded vs multithreaded library generation.

+++ libraries/liblutil/passwd.c
@@ -81,7 +81,7 @@ lutil_passwd(
 		unsigned char *orig_pass = NULL;
  
 		/* base64 un-encode password */
-		orig_pass = (unsigned char *)malloc(pw_len * 0.75 + 1);
+		orig_pass = (unsigned char *)malloc((size_t)(pw_len * 0.75 + 1));
 		if ((rc = lutil_b64_pton(p, orig_pass, pw_len)) < 0)
 		{
 			free(orig_pass);
@@ -111,7 +111,7 @@ lutil_passwd(
 		unsigned char *orig_pass = NULL;
 
 		/* base64 un-encode password */
-		orig_pass = (unsigned char *)malloc(pw_len * 0.75 + 1);
+		orig_pass = (unsigned char *)malloc((size_t)(pw_len * 0.75 + 1));
 		if ((rc = lutil_b64_pton(p, orig_pass, pw_len)) < 0)
 		{
 			free(orig_pass);
